+++ packages/core/dropdown-menu/examples/08-with-disabled-button-trigger.js
@@ -1,13 +1,14 @@
 // @flow
 
 import React from 'react';
+import { defaultProps as defaultButtonProps } from '@atlaskit/button';
 import DropdownMenu, { DropdownItemGroup, DropdownItem } from '../src';
 
 export default () => (
   <DropdownMenu
     trigger="Disabled trigger"
     triggerType="button"
-    triggerButtonProps={{ isDisabled: true }}
+    triggerButtonProps={{ ...defaultButtonProps, isDisabled: true }}
   >
     <DropdownItemGroup>
       <DropdownItem>Sydney</DropdownItem>