+++ packages/search/global-search/src/components/confluence/ConfluenceQuickSearchContainer.tsx
@@ -12,6 +12,7 @@ import { PeopleSearchClient } from '../../api/PeopleSearchClient';
 import renderSearchResults from './ConfluenceSearchResults';
 import settlePromises from '../../util/settle-promises';
 import { LinkComponent } from '../GlobalQuickSearchWrapper';
+import { redirectToConfluenceAdvancedSearch } from '../SearchResultsUtil';
 
 export interface Props {
   crossProductSearchClient: CrossProductSearchClient;
@@ -74,6 +75,11 @@ export class ConfluenceQuickSearchContainer extends React.Component<
     } else {
       this.doSearch(query);
     }
+  };
+
+  handleSearchSubmit = () => {
+    const { query } = this.state;
+    redirectToConfluenceAdvancedSearch(query);
   };
 
   async searchCrossProductConfluence(
@@ -200,6 +206,7 @@ export class ConfluenceQuickSearchContainer extends React.Component<
       <GlobalQuickSearch
         onMount={this.handleMount}
         onSearch={this.handleSearch}
+        onSearchSubmit={this.handleSearchSubmit}
         isLoading={isLoading}
         query={query}
         linkComponent={linkComponent}